% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.addChangedAttr,simList-method}
\alias{.addChangedAttr,simList-method}
\title{\code{.addChangedAttr} for \code{simList} objects}
\usage{
\S4method{.addChangedAttr}{simList}(object, preDigest, origArguments, ...)
}
\value{
returns the object with attribute added
}
\description{
This will evaluate which elements in the \code{simList} object changed following
this Cached function call. It will add a named character string as an
attribute \code{attr(x, ".Cache")$changed}, indicating which ones changed.
When this function is subsequently called again, only these changed objects
will be returned. All other \code{simList} objects will remain unchanged.
}
\seealso{
\link[reproducible:.addChangedAttr]{reproducible::.addChangedAttr}
}
