% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simList-accessors.R
\name{.addDepends}
\alias{.addDepends}
\alias{.addDepends,simList,.moduleDeps-method}
\title{Add simulation dependencies}
\usage{
.addDepends(sim, x)

\S4method{.addDepends}{simList,.moduleDeps}(sim, x)
}
\arguments{
\item{sim}{A \code{simList} object from which to extract element(s) or in which to replace element(s).}

\item{x}{A named list containing the parameters used to construct a new
\code{.moduleDeps} object.}
}
\value{
A \code{simList} object.
}
\description{
Internal function.
Adds a \code{.moduleDeps} object to the simulation dependency list.
}
\seealso{
Other functions to access elements of a 'simList' object:
\code{\link[=checkpointFile]{checkpointFile()}},
\code{\link[=envir]{envir()}},
\code{\link[=events]{events()}},
\code{\link[=globals]{globals()}},
\code{\link[=inputs]{inputs()}},
\code{\link[=modules]{modules()}},
\code{\link[=objs]{objs()}},
\code{\link[=packages]{packages()}},
\code{\link[=params]{params()}},
\code{\link[=paths]{paths()}},
\code{\link[=progressInterval]{progressInterval()}},
\code{\link[=times]{times()}}
}
\author{
Alex Chubaty
}
\concept{functions to access elements of a 'simList' object}
\keyword{internal}
