% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.addTagsToOutput,simList-method}
\alias{.addTagsToOutput,simList-method}
\title{\code{.addTagsToOutput} for \code{simList} objects}
\usage{
\S4method{.addTagsToOutput}{simList}(object, outputObjects, FUN, preDigestByClass)
}
\value{
modified \code{object}, with attributes added
}
\description{
See \code{\link[reproducible:.addTagsToOutput]{reproducible::.addTagsToOutput()}}.
}
\author{
Eliot McIntire
}
