% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{all.equal.simList}
\alias{all.equal.simList}
\title{All equal method for \code{simList} objects}
\usage{
\method{all}{equal.simList}(target, current, ...)
}
\arguments{
\item{target}{\R object.}

\item{current}{other \R object, to be compared with \code{target}.}

\item{...}{further arguments for different methods, notably the
    following two, for numerical comparison:}
}
\value{
See \code{\link[base:all.equal]{base::all.equal()}}
}
\description{
This function removes a few attributes that are added internally
by \pkg{SpaDES.core} and are not relevant to the \code{all.equal}.
One key element removed is any time stamps, as these are guaranteed to be different.
A possibly very important argument to pass to the \code{...} is \code{check.attributes = FALSE}
which will allow successful comparisons of many objects that might have pointers.
}
