% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-methods.R
\name{bindrows}
\alias{bindrows}
\title{Simple wrapper around \code{data.table::rbindlist}}
\usage{
bindrows(...)
}
\description{
This simply sets defaults to \code{fill = TRUE}, and
\code{use.names = TRUE}
}
