% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-template.R
\name{checkModulePath}
\alias{checkModulePath}
\title{Uses "." if getPath not set}
\usage{
checkModulePath()
}
\description{
Will compare default in spadesOptions to getPaths ... these will be same if use
has not set them. For such case, use ".". They will be different if the user has
used \code{setPaths}. If that is the case, then use \code{getPaths()[["modulePath"]]}
}
