% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadData.R
\name{checksums}
\alias{checksums}
\title{Calculate checksum  or a module's data files}
\usage{
checksums(module, path, ...)
}
\arguments{
\item{module}{Character string giving the name of the module.}

\item{path}{Character string giving the path to the module directory.}

\item{...}{Passed to \code{\link[reproducible]{Checksums}}, notably, \code{write},
\code{quickCheck},  \code{checksumFile} and \code{files}.}
}
\description{
Verify (and optionally write) checksums for data files in a module's
\code{data/} subdirectory. The file \code{data/CHECKSUMS.txt} contains the
expected checksums for each data file.
Checksums are computed using \code{SpaDES.tools:::.digest}, which is simply a
wrapper around \code{digest::digest}.
}
\details{
Modules may require data that for various reasons cannot be distributed with
the module source code. In these cases, the module developer should ensure
that the module downloads and extracts the data required. It is useful to not
only check that the data files exist locally but that their checksums match
those expected.
}
\note{
In version 1.2.0 and earlier, two checksums per file were required
because of differences in the checksum hash values on Windows and Unix-like
platforms. Recent versions use a different (faster) algorithm and only require
one checksum value per file.
To update your \file{CHECKSUMS.txt} files using the new algorithm, see
\url{https://github.com/PredictiveEcology/SpaDES/issues/295#issuecomment-246513405}.
}
