% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\docType{methods}
\name{clearCache,simList-method}
\alias{clearCache,simList-method}
\title{clearCache for simList objects}
\usage{
\S4method{clearCache}{simList}(x, userTags = character(), after, before,
  ask = getOption("reproducible.ask"), useCloud = FALSE,
  cloudFolderID = getOption("reproducible.cloudFolderID", NULL), ...)
}
\description{
This will take the \code{cachePath(object)} and pass
}
