% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadInputs.R
\name{cropReprojInputs}
\alias{cropReprojInputs}
\title{Reproject, crop module inputs}
\usage{
cropReprojInputs(x, studyArea = NULL, rasterToMatch = NULL,
  rasterInterpMethod = "bilinear", addTagsByObject = NULL,
  cacheTags = NULL)
}
\arguments{
\item{x}{A \code{Spatial*}, \code{sf}, or \code{Raster*} object.}

\item{studyArea}{Template \code{SpatialPolygons*} object used for masking, after cropping.
If not in same CRS, then it will be \code{spTransform}ed to
CRS of \code{x} before masking.}

\item{rasterToMatch}{Template \code{Raster*} object used for reprojecting and
cropping.}

\item{rasterInterpMethod}{Method used to compute values for the new
\code{RasterLayer}. See \code{\link[raster]{projectRaster}}. Defaults to bilinear.}

\item{addTagsByObject}{Pass any object in there for which there is a
\code{.tagsByClass} function}

\item{cacheTags}{Character vector with Tags. These Tags will be added to the
repository along with the artifact.}
}
\description{
This function can be used to crop or reproject module inputs from raw data.
}
\author{
Eliot McIntire

Jean Marchal
}
