% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveLoadSimList.R, R/spades-core-deprecated.R
\name{zipSimList}
\alias{zipSimList}
\alias{experiment}
\alias{experiment2}
\alias{POM}
\alias{simInitAndExperiment}
\alias{loadPackages}
\title{Zip a \code{simList} and various files}
\usage{
zipSimList(sim, zipfile, ..., outputs = TRUE, inputs = TRUE, cache = FALSE)

experiment(...)

experiment2(...)

POM(...)

simInitAndExperiment(...)

loadPackages(...)
}
\arguments{
\item{sim}{Either a \code{simList} or a character string of the name
of a \code{simList} that can be found in \code{envir}.
Using a character string will assign that object name to the saved
\code{simList}, so when it is recovered it will be given that name.}

\item{zipfile}{A character string indicating the filename for the zip file. Passed to \code{zip}.}

\item{...}{Unused.}

\item{outputs}{Logical. If \code{TRUE}, all files identified in
\code{outputs(sim)} will be included in the zip.}

\item{inputs}{Logical. If \code{TRUE}, all files identified in
\code{inputs(sim)} will be included in the zip.}

\item{cache}{Logical. Not yet implemented. If \code{TRUE}, all files in \code{cachePath(sim)}
will be included in the archive.
Defaults to \code{FALSE} as this could be large, and may include many out of date elements.
See Details.}
}
\value{
These functions are defunct here and are called for their side effect of
signalling a deprecation message pointing at the package that now provides them.
They return \code{NULL}, invisibly.
}
\description{
\code{zipSimList} will save the \code{simList} and file-backed \verb{Raster*} objects, plus,
optionally, files identified in \code{outputs(sim)} and \code{inputs(sim)}.
This uses \code{Copy} under the hood, to not affect the original \code{simList}.

These functions have been moved to other packages.
}
