% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-repository.R
\docType{methods}
\name{.digest}
\alias{.digest}
\alias{.digest,character-method}
\title{Calculate the hashes of multiple files}
\usage{
.digest(file, quickCheck, ...)

\S4method{.digest}{character}(file, quickCheck, algo = "xxhash64", ...)
}
\arguments{
\item{file}{Character vector of file paths.}

\item{quickCheck}{Logical. If \code{TRUE}, then the check with local data will only
use \code{file.size} instead of \code{digest::digest}.
This is faster, but potentially much less robust.}

\item{...}{Additional arguments to \code{digest::digest}.}
}
\value{
A character vector of hashes.
}
\description{
Internal function. Wrapper for \code{\link[digest]{digest}} using md5sum.
}
\author{
Alex Chubaty
}
\keyword{internal}
