% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{.guessPkgFun}
\alias{.guessPkgFun}
\title{Guess package of a function}
\usage{
.guessPkgFun(bsf)
}
\arguments{
\item{bsf}{character. A function name}
}
\value{
character. The package and function name as \code{"pkg::bsf"}
}
\description{
Guess package of a function
}
