% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-parseModule.R
\name{evalWithActiveCode}
\alias{evalWithActiveCode}
\title{Evaluate code that has "active" code, i.e., functions like start(sim)}
\usage{
evalWithActiveCode(parsedModuleNoDefineModule, envir,
  parentFrame = parent.frame())
}
\description{
Evaluate code that has "active" code, i.e., functions like start(sim)
}
