% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadInputs.R
\name{extractFromArchive}
\alias{extractFromArchive}
\title{Extract files from archive.}
\usage{
extractFromArchive(archive, destinationPath = dirname(archive), needed,
  extractedArchives = NULL)
}
\arguments{
\item{archive}{Character string giving the path of the archive
containing the \code{file} to be extracted.}

\item{destinationPath}{Character string giving the path where \code{needed} will be
extracted. Defaults to the archive directory.}

\item{needed}{Character string giving the name of the file(s) to be extracted.}

\item{extractedArchives}{Used internally.}
}
\value{
A character vector listing the paths of the extracted archives.
}
\description{
Extract zip or tar archive files, possibly nested in other zip or tar
archives.
}
\author{
Jean Marchal
}
