% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadData.R
\docType{methods}
\name{extractURL,missing-method}
\alias{extractURL,missing-method}
\alias{extractURL,simList-method}
\alias{extractURL,NULL-method}
\title{Extract a URL}
\usage{
\S4method{extractURL}{missing}(x)

\S4method{extractURL}{simList}(x)

\S4method{extractURL}{`NULL`}(x)
}
\arguments{
\item{x}{A character string or other object containing a url.}
}
\description{
Extract a URL
}
\author{
Eliot McIntire
}
