% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-template.R
\name{.fileEdit}
\alias{.fileEdit}
\title{Open a file for editing}
\usage{
.fileEdit(file)
}
\arguments{
\item{file}{Character string giving the file path to open.}
}
\value{
Invoked for its side effect of opening a file for editing.
}
\description{
RStudio's \code{file.edit} behaves differently than \code{utils::file.edit}.
The workaround is to have the user manually open the file if they are using RStudio.
}
\author{
Alex Chubaty
}
\keyword{internal}
