% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-methods.R
\name{.fileTableIn}
\alias{.fileTableIn}
\alias{.fileTableIn,missing-method}
\alias{.fileTableInCols}
\alias{.fileTableInDF}
\alias{.fileTableOut}
\alias{.fileTableOut,missing-method}
\alias{.fileTableOutCols}
\alias{.fileTableOutDF}
\title{Create empty \code{fileTable} for inputs and outputs}
\usage{
.fileTableIn(x)

\S4method{.fileTableIn}{missing}()

.fileTableInCols

.fileTableInDF

.fileTableOut(x)

\S4method{.fileTableOut}{missing}()

.fileTableOutCols

.fileTableOutDF
}
\arguments{
\item{x}{Not used (should be missing)}
}
\value{
An empty data.frame with structure needed for input/output \code{fileTable.}
}
\description{
Internal functions.
Returns an empty \code{fileTable} to be used with inputs and outputs.
}
\keyword{internal}
