% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-define.R
\name{.fillOutputRows}
\alias{.fillOutputRows}
\title{An internal function for coercing a data.frame to \code{outputs()}}
\usage{
.fillOutputRows(outputDF, endTime)
}
\arguments{
\item{outputDF}{A data.frame with partial columns to pass to \verb{outputs<-}}

\item{endTime}{Numeric time. The \code{end(sim)}.}
}
\description{
An internal function for coercing a data.frame to \code{outputs()}
}
\keyword{internal}
