% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-methods.R
\name{FilterRecursive}
\alias{FilterRecursive}
\title{Recursively Filter a list}
\usage{
FilterRecursive(f, x)
}
\arguments{
\item{f}{a unary function}

\item{x}{a vector}
}
\description{
Recursively Filter a list
}
\keyword{internal}
