% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-checking.R
\name{.findElementsInEnv}
\alias{.findElementsInEnv}
\alias{.findElement}
\title{Find all references to sim$}
\usage{
.findElementsInEnv(envToFindSim, moduleEnv, type)

.findElement(x, type)
}
\arguments{
\item{envToFindSim}{An environment where sim is defined}

\item{moduleEnv}{The environment where the module functions are}

\item{type}{Either "get", "assign", or "globals". See details.}

\item{x}{A call in which to search for sim}
}
\value{
A character string with all sim objects found
}
\description{
Find all references to sim$
}
\details{
When \bold{\code{type = "get"}}, the function scans for sim$xxx or sim[["xxx"]]] on
the RHS of an assignment operator or when there is no assignment. When
\bold{\code{type = "assign"}}, the function scans for sim$xxx or sim[['xxx']] on the
LHS of an assignment operator. When \bold{\code{type = "globals"}}, the function
scans for all functions (i.e., "globals") being used. This is similar to
\code{\link[codetools]{findGlobals}}, but faster.

\code{.findElement} will omit whatever it finds inside a \code{is.null}, when
\code{type = "assign"}. Usually this is a test of existence of that object, in
order to assign to that object.
}
\author{
Eliot McIntire
}
\keyword{internal}
