% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.findSimList}
\alias{.findSimList}
\title{Find simList in a nested list}
\usage{
.findSimList(x)
}
\arguments{
\item{x}{any object, used here only when it is a list with at least one
\code{simList} in it}
}
\description{
THis is recursive, so it will find the all simLists even if they are deeply nested.
}
