% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadInputs.R
\name{fixErrors}
\alias{fixErrors}
\title{Do some minor error fixing}
\usage{
fixErrors(x, targetFile, attemptErrorFixes = TRUE, ...)
}
\arguments{
\item{targetFile}{Character string giving the path to the eventual
file (raster, shapefile, csv, etc.) after downloading and extracting from
a zip or tar archive. This is the file \emph{before}
it is passed to \code{postProcess}. Currently, the internal
checksumming does not checksum the file after it is
\code{postProcess}ed (e.g., cropped/reprojected/masked).
Using \code{Cache} around \code{prepInputs} will do a
sufficient job in these cases.}

\item{attemptErrorFixes}{Will attempt to fix known errors. Currently only some failures
for SpatialPolygons* are attempted. Notably with \code{raster::buffer(..., width = 0)}.
Default \code{TRUE}, though this may not be the right action for all cases.}

\item{...}{None used currently}
}
\description{
These must be very common for this function to be useful. Currently, the only
meaningful method is on SpatialPolygons, and it runs \code{rgeos::gIsValid}. If
\code{FALSE}, then it runs a buffer of width 0.
}
