% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadInputs.R
\name{fixErrors.SpatialPolygons}
\alias{fixErrors.SpatialPolygons}
\title{Fix \code{rgeos::gIsValid} failures in \code{SpatialPolygons}}
\usage{
\method{fixErrors}{SpatialPolygons}(x, targetFile, attemptErrorFixes = TRUE,
  ...)
}
\arguments{
\item{x}{A \code{SpatialPolygons} object}

\item{targetFile}{Character string giving the path to the eventual
file (raster, shapefile, csv, etc.) after downloading and extracting from
a zip or tar archive. This is the file \emph{before}
it is passed to \code{postProcess}. Currently, the internal
checksumming does not checksum the file after it is
\code{postProcess}ed (e.g., cropped/reprojected/masked).
Using \code{Cache} around \code{prepInputs} will do a
sufficient job in these cases.}

\item{attemptErrorFixes}{Will attempt to fix known errors. Currently only some failures
      for SpatialPolygons* are attempted. Notably with \code{raster::buffer(..., width = 0)}.
      Default \code{TRUE}, though this may not be the right action for all cases.
@examples}

\item{...}{None used currently}
}
\description{
This uses \code{raster::buffer(..., width = 0)} internally, which fixes some
failures to \code{rgeos::gIsValid}
}
