% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-parseModule.R
\name{.getModuleInputObjects}
\alias{.getModuleInputObjects}
\title{Extract the user-defined \code{.inputObjects} function from a module}
\usage{
.getModuleInputObjects(sim, m)
}
\description{
Extract the user-defined \code{.inputObjects} function from a module
}
\keyword{internal}
