% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getMapPath}
\alias{getMapPath}
\alias{getSampleModules}
\title{Get copies of sample files for examples and tests}
\usage{
getMapPath(tmpdir)

getSampleModules(tmpdir)
}
\arguments{
\item{tmpdir}{character specifying the path to a temporary directory (e.g., \code{tempdir()})}
}
\value{
character vector of filepaths to the copied files
}
\description{
Get copies of sample files for examples and tests
}
