% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadInputs.R
\name{guessAtTarget}
\alias{guessAtTarget}
\alias{.guessAtTarget}
\title{Try to pick a file to load}
\usage{
.guessAtTarget(targetFilePath, destinationPath, fun)
}
\arguments{
\item{targetFilePath}{Full path of the target file}

\item{destinationPath}{Full path of the directory where the target file should be}
}
\description{
Try to pick a file to load
}
\keyword{internal}
