% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-dependencies-class.R
\name{._inputObjectsDF}
\alias{._inputObjectsDF}
\alias{._inputObjectsDF,missing-method}
\alias{._outputObjectsDF}
\alias{._outputObjectsDF,missing-method}
\title{Create an empty \code{data.frame} object for use with \code{inputObjects} or
\code{outputObjects}}
\usage{
._inputObjectsDF(x)

\S4method{._inputObjectsDF}{missing}()

._outputObjectsDF(x)

\S4method{._outputObjectsDF}{missing}()
}
\arguments{
\item{x}{Not used. Should be missing.}
}
\value{
A \code{data.frame} object.
}
\description{
Internal function.
}
\seealso{
\code{\link{defineModule}}
}
\author{
Alex Chubaty
}
\keyword{internal}
