% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-parseModule.R
\name{.isNamespaced}
\alias{.isNamespaced}
\title{Check is module uses module namespacing}
\usage{
.isNamespaced(sim, m)
}
\description{
Older modules may not have their functions etc. namespaced in the \code{simList}.
}
\keyword{internal}
