% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.keepAttrs}
\alias{.keepAttrs}
\title{Attach missing attributes from x to y}
\usage{
.keepAttrs(x, y, omitAttrs = c(".envir", ".list", ".xData", ".Data"))
}
\arguments{
\item{x}{an object with attributes}

\item{y}{an object with attributes}
}
\description{
This is an internal helper.
}
\keyword{internal}
