% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{makeMemoiseable.simList}
\alias{makeMemoiseable.simList}
\alias{unmakeMemoiseable.simList_}
\title{Make simList correctly work with memoise}
\usage{
\method{makeMemoiseable}{simList}(x)

\method{unmakeMemoiseable}{simList_}(x)
}
\arguments{
\item{x}{An object to make memoiseable.
See individual methods in other packages.}
}
\value{
A \code{simList_} object or a \code{simList}, in the case
of \code{unmakeMemoiseable}.
}
\description{
Because of the environment slot, \code{simList} objects don't
correctly memoise a \code{simList}. This method for
\code{simList} converts the object to a \code{simList_} first.
}
\seealso{
\code{\link[reproducible]{makeMemoiseable}}
}
