% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadInputs.R
\name{maskInputs}
\alias{maskInputs}
\title{Mask module inputs}
\usage{
maskInputs(x, studyArea, ...)
}
\arguments{
\item{x}{A \code{Raster*} object}

\item{studyArea}{A \code{SpatialPolygons*} object}
}
\description{
This function can be used to mask module inputs from raw data.
}
\author{
Eliot McIntire

Jean Marchal
}
