% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{memoryUseThisSession}
\alias{memoryUseThisSession}
\title{Estimate memory used with \code{system("ps")}}
\usage{
memoryUseThisSession(thisPid)
}
\description{
This will give a slightly different estimate than \code{pryr::mem_used},
which uses \code{gc()} internally. The purpose of this function is
to allow continuous monitoring, external to the R session. Normally,
this is run in a different session.
}
