% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-methods.R
\name{moduleCodeFiles}
\alias{moduleCodeFiles}
\title{Extract the full file paths for R source code}
\usage{
moduleCodeFiles(paths, modules)
}
\arguments{
\item{paths}{An optional named list with up to 4 named elements,
\code{modulePath}, \code{inputPath}, \code{outputPath}, and \code{cachePath}.
See details. NOTE: Experimental feature now allows for multiple \code{modulePath}s
to be specified in a character vector. The modules will be searched for sequentially
in the first \code{modulePath}, then if it doesn't find it, in the second etc.}

\item{modules}{A named list of character strings specifying the names of modules to be loaded
for the simulation.
Note: the module name should correspond to the R source file from which the module is loaded.
Example: a module named "caribou" will be sourced form the file \file{caribou.R},
located at the specified \code{modulePath(simList)} (see below).}
}
\description{
This can be used e.g., for Caching, to identify which files have changed.
}
