% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-define.R
\name{moduleDefaults}
\alias{moduleDefaults}
\title{Defaults values used in \code{defineModule}}
\usage{
moduleDefaults
}
\value{
named list of default module metadata
}
\description{
Where individual elements are missing in \code{defineModule}, these defaults will be used.
}
