% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simList-accessors.R
\name{.callingModuleName}
\alias{.callingModuleName}
\alias{.callingModuleName,simList-method}
\alias{currentModule}
\alias{currentModule,simList-method}
\alias{.callingFrameTimeunit}
\title{Namespacing within \code{SpaDES}}
\usage{
.callingModuleName(sim)

\S4method{.callingModuleName}{simList}(sim)

currentModule(sim)

\S4method{currentModule}{simList}(sim)

.callingFrameTimeunit(x)
}
\arguments{
\item{sim}{A \code{simList} object from which to extract element(s) or
in which to replace element(s).}

\item{x}{A \code{simList}}
}
\value{
\code{.callingModuleName} returns the name of the module that is currently
the active module calling functions like \code{scheduleEvent}.
}
\description{
This will only return the module name if it is inside a \code{spades} call,
i.e., it will return \code{NULL} if used in interactive mode.
The related function \code{currentModule} is simply a rapid accessor for the
current module name. This latter will return the module that is in the current
event queue, which will never be \code{NULL}.
}
\author{
Eliot McIntire
}
\keyword{internal}
