% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-template.R
\docType{methods}
\name{newProject}
\alias{newProject}
\alias{newProject,character,character-method}
\title{Create new SpaDES project}
\usage{
newProject(name, path, ...)

\S4method{newProject}{character,character}(name, path, ...)
}
\arguments{
\item{name}{project name (name of project directory)}

\item{path}{path to directory in which to create the project directory}

\item{...}{additional arguments (not used)}
}
\description{
Initialize a project with subdirectories \file{cache/}, \file{modules/},
\file{inputs/}, \file{outputs/}, and \code{setPaths} accordingly.
}
\examples{
myProjDir <- newProject("myProject", tempdir())

dir.exists(file.path(myProjDir, "cache"))
dir.exists(file.path(myProjDir, "inputs"))
dir.exists(file.path(myProjDir, "modules"))
dir.exists(file.path(myProjDir, "outputs"))
unlink(myProjDir, recursive = TRUE) ## cleanup
}
