% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{noEventWarning}
\alias{noEventWarning}
\title{Text for no event with that name}
\usage{
noEventWarning(sim)
}
\arguments{
\item{sim}{A \code{simList} simulation object, generally produced by \code{simInit}.}
}
\value{
A text string specifying the event name and module for which there is no event
}
\description{
Provides the text to be sent to \code{warning} in each module as the default \code{switch} case.
}
