% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-diagrams.R
\name{objectDiagram}
\alias{objectDiagram}
\alias{objectDiagram,simList-method}
\title{Simulation object dependency diagram}
\usage{
objectDiagram(sim, ...)

\S4method{objectDiagram}{simList}(sim, ...)
}
\arguments{
\item{sim}{A \code{simList} object (typically corresponding to a
completed simulation).}

\item{...}{Additional arguments passed to \code{DiagrammeR::mermaid}.
Useful for specifying \code{height} and \code{width}.}
}
\value{
Plots a sequence diagram, invisibly returning a
  \code{DiagrammeR::mermaid} object.
}
\description{
Create a sequence diagram illustrating the data object dependencies of a
simulation. Offers a more detailed view of specific objects than does
plotting the \code{depsEdgeList} directly with \code{\link{moduleDiagram}}.
}
\examples{
\dontrun{
objectDiagram(sim)
# if there are lots of objects, may need to increase width and/or height
objectDiagram(sim, height = 3000, width = 3000)
}

}
\seealso{
\code{DiagrammeR::mermaid}.
}
\author{
Alex Chubaty
}
