% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{.objectNames}
\alias{.objectNames}
\title{Extract object names}
\usage{
.objectNames(calledFrom = "spades", argClass = "simList", argName = "")
}
\arguments{
\item{calledFrom}{character vector of length 1, indicating which function
call is desired. Defaults to \code{"spades"}.}

\item{argClass}{character vector of length 1, indicating which class is
being searched for among the arguments.
Defaults to \code{"simList"}.}

\item{argName}{character vector of length 1, or \code{NULL}, indicating
if the arguments to select have a name, no name (empty
string), or do not use name (\code{NULL}).}
}
\value{
\code{NULL}. This function is invoked for its side effects.
}
\description{
Extract object names
}
\seealso{
Other internals from \code{quickPlot}: 
\code{\link{.parseArgs}()},
\code{\link{.parseElems}()},
\code{\link{whereInStack}()}
}
\author{
Eliot McIntire
}
\concept{internals from \code{quickPlot}}
\keyword{internal}
