% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simList-accessors.R
\name{packages}
\alias{packages}
\alias{simList-accessors-packages}
\alias{packages,ANY-method}
\title{Get module or simulation package dependencies}
\usage{
packages(sim, modules, paths, filenames, envir, clean = FALSE, ...)

\S4method{packages}{ANY}(sim, modules, paths, filenames, envir, clean = FALSE, ...)
}
\arguments{
\item{sim}{A \code{simList} object.}

\item{modules}{Character vector, specifying the name or vector of names of module(s)}

\item{paths}{Character vector, specifying the name or vector of names of paths(s) for
those modules. If path not specified, it will be taken from
\code{getOption("spades.modulePath")}, which is set with \code{setPaths()})}

\item{filenames}{Character vector specifying filenames of modules (i.e.
combined path & module. If this is specified, then \code{modules} and
\code{path} are ignored.}

\item{envir}{Optional environment in which to store parsed code. This may be
useful if the same file is being parsed multiple times. This
function will check in that environment for the parsed file before
parsing again. If the \code{envir} is transient, then this will have no effect.}

\item{clean}{Optional logical. If \code{TRUE}, it will scrub any references to
GitHub repositories, e.g., "PredictiveEcology/reproducible" will be
returned as "reproducible".}

\item{...}{All \code{simInit} parameters.}
}
\value{
A sorted character vector of package names.
}
\description{
Get module or simulation package dependencies
}
\seealso{
Other functions to access elements of a 'simList' object:
\code{\link[=.addDepends]{.addDepends()}},
\code{\link[=checkpointFile]{checkpointFile()}},
\code{\link[=envir]{envir()}},
\code{\link[=events]{events()}},
\code{\link[=globals]{globals()}},
\code{\link[=inputs]{inputs()}},
\code{\link[=modules]{modules()}},
\code{\link[=objs]{objs()}},
\code{\link[=params]{params()}},
\code{\link[=paths]{paths()}},
\code{\link[=progressInterval]{progressInterval()}},
\code{\link[=times]{times()}}
}
\author{
Alex Chubaty & Eliot McIntire
}
\concept{functions to access elements of a 'simList' object}
