% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{.parseElems}
\alias{.parseElems}
\alias{.parseElems,ANY-method}
\title{Parsing of elements}
\usage{
.parseElems(tmp, elems, envir)

\S4method{.parseElems}{ANY}(tmp, elems, envir)
}
\arguments{
\item{tmp}{A evaluated object}

\item{elems}{A character string to be parsed}

\item{envir}{An environment}
}
\value{
An object, parsed from a character string and and environment
}
\description{
This is a generic definition that can be extended according to class.
Intended only for development use.
}
\seealso{
Other internals from \code{quickPlot}: 
\code{\link{.objectNames}()},
\code{\link{.parseArgs}()},
\code{\link{whereInStack}()}
}
\author{
Eliot McIntire
}
\concept{internals from \code{quickPlot}}
\keyword{internal}
