% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-parseModule.R
\name{.parseMessage}
\alias{.parseMessage}
\title{Prepend module name to a message}
\usage{
.parseMessage(m, message)
}
\arguments{
\item{m}{module name}

\item{message}{rest of message}
}
\value{
A message with that starts with paste0(m, ":", message)
}
\description{
Also makes it blue.
}
\keyword{internal}
