% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.preDigestByClass,simList-method}
\alias{.preDigestByClass,simList-method}
\title{Pre-digesting method for \code{simList}}
\usage{
\S4method{.preDigestByClass}{simList}(object)
}
\value{
character vector corresponding to the names of objects stored in the \code{.xData} slot
}
\description{
Takes a snapshot of \code{simList} objects.
}
\details{
See \code{\link[reproducible:.preDigestByClass]{reproducible::.preDigestByClass()}}.
}
\seealso{
\link[reproducible:.preDigestByClass]{reproducible::.preDigestByClass}
}
\author{
Eliot McIntire
}
