% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.prepareOutput,simList-method}
\alias{.prepareOutput,simList-method}
\title{\code{.prepareOutput} for \code{simList} objects}
\usage{
\S4method{.prepareOutput}{simList}(object, cachePath, ...)
}
\value{
the modified \code{object}
}
\description{
See \code{\link[reproducible:.prepareOutput]{reproducible::.prepareOutput()}}.
}
\seealso{
\link[reproducible:.prepareOutput]{reproducible::.prepareOutput}
}
