% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{newProgressBar}
\alias{newProgressBar}
\alias{setProgressBar}
\title{Progress bar}
\usage{
newProgressBar(sim)

setProgressBar(sim)
}
\arguments{
\item{sim}{A \code{simList} simulation object.}
}
\description{
Shows a progress bar that is scaled to simulation end time.
}
\details{
The progress bar object is stored in a separate environment, #' \code{.pkgEnv}.
}
\author{
Alex Chubaty and Eliot McIntire
}
