% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{.quickCheck}
\alias{.quickCheck}
\title{The \code{SpaDES.core} variable to switch between quick and robust checking}
\usage{
.quickCheck
}
\description{
A variable that can be use by module developers and model users to switch between
a quick check of functions like \code{downloadData}, \code{Cache.}
The module developer must actually use this in their code.
}
