% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{rasterCreate}
\alias{rasterCreate}
\alias{rasterCreate.default}
\title{Simple wrapper to load any \verb{Raster*} object}
\usage{
rasterCreate(x, ...)

\method{rasterCreate}{default}(x, ...)
}
\arguments{
\item{x}{An object, notably a \verb{Raster*} object. All others will simply
be passed through with no effect.}

\item{...}{Passed to \code{raster::raster}, \code{raster::stack},
or \code{raster::brick}}
}
\value{
a new (empty) object of same class as the original.
}
\description{
This wraps either \code{raster::raster}, \code{raster::stack}, \code{raster::brick}, or \code{terra::rast},
allowing a single function to be used to create a new object of the same class as a template.
This works for all \verb{Raster*} and \code{SpatRaster} class templates.
}
\section{Methods (by class)}{
\itemize{
\item \code{rasterCreate(default)}: Simply passes through argument with no effect

}}
