% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{rasterToMemory}
\alias{rasterToMemory}
\alias{rasterToMemory,Raster-method}
\alias{rasterToMemory,list-method}
\alias{rasterToMemory,ANY-method}
\alias{rasterToMemory,simList-method}
\title{Read raster to memory}
\usage{
rasterToMemory(x, ...)

\S4method{rasterToMemory}{Raster}(x, ...)

\S4method{rasterToMemory}{list}(x, ...)

\S4method{rasterToMemory}{ANY}(x, ...)

\S4method{rasterToMemory}{simList}(x, ...)
}
\arguments{
\item{x}{An object passed directly to the function raster (e.g., character string of a filename).}

\item{...}{Additional arguments to \code{raster::raster}, \code{raster::stack}, or \code{raster::brick}.}
}
\value{
A raster object whose values are stored in memory.
}
\description{
Wrapper to the \code{raster} function, that creates the raster object in
memory, even if it was read in from file. There is the default method which is
just a pass through, so this can be safely used on large complex objects,
recursively, e.g., a \code{simList}.
}
\seealso{
\code{\link[=raster]{raster()}}.
}
\author{
Eliot McIntire and Alex Chubaty
}
