% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{paddedFloatToChar}
\alias{padYears}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{reproducible}{\code{\link[reproducible:paddedFloatToChar]{paddedFloatToChar()}}, \code{\link[reproducible:padYears]{padYears()}}}
}}

