% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadData.R
\name{remoteFileSize}
\alias{remoteFileSize}
\title{Determine the size of a remotely hosted file}
\usage{
remoteFileSize(url)
}
\arguments{
\item{url}{The url of the remote file.}
}
\value{
A numeric indicating the size of the remote file in bytes.
}
\description{
Defunct. Will be removed by mid-2023.
}
\author{
Eliot McIntire and Alex Chubaty
}
