% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python-ReticulateFindPython.R
\name{reticulate_python_path}
\alias{reticulate_python_path}
\title{Reticulate: Return Python interpreter path}
\usage{
reticulate_python_path(version = NULL, pyenvRoot = NULL, pyenvOnly = FALSE)
}
\arguments{
\item{version}{character. Python version or a comma separated list of version constraints.
See \code{\link[reticulate]{virtualenv_starter}} for more details.}

\item{pyenvRoot}{character. Path to directory containing pyenv-win tool.}

\item{pyenvOnly}{logical. Exclude versions not installed by pyenv.}
}
\value{
character or NULL. If found, a path to Python interpreter.
}
\description{
Get path to a Python interpreter including installs at a given pyenv-win location.
}
\keyword{internal}
