% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python-ReticulateFindPython.R
\name{ReticulateFindPython}
\alias{ReticulateFindPython}
\title{Reticulate Find Python}
\usage{
ReticulateFindPython(
  version,
  versionInstall = version,
  pyenvRoot = NULL,
  pyenvOnly = FALSE,
  useGit = TRUE,
  prompt = FALSE
)
}
\arguments{
\item{version}{character. Python version or a comma separated list of version constraints.
See \code{\link[reticulate]{virtualenv_starter}} for more details.}

\item{versionInstall}{character. Version to install if suitable version not found.
Required if \code{version} is a list of version constraints.}

\item{pyenvRoot}{character. Path of location for install of the pyenv-win tool.
Defaults to the R user data directory.}

\item{pyenvOnly}{logical. Exclude versions not within a pyenv install directory.}

\item{useGit}{logical. Allow \code{reticulate} to clone pyenv-win if Git is available.}

\item{prompt}{logical. Prompt user to approve download of pyenv-win tool.}
}
\value{
character. Path to Python interpreter.
}
\description{
Use \code{reticulate} to find or install Python to meet version requirements.
Download and use the \href{https://pypi.org/project/pyenv-win/}{pyenv-win}
Python version management tool from Github if necessary.
}
\details{
This function was created to bypass the requirement for Git to be installed
when \code{\link[reticulate]{install_python}} is called on a Windows computer
without pyenv-win already installed.
If Git is available, \code{reticulate} will clone the
\href{https://github.com/pyenv-win/pyenv-win}{pyenv-win Github repository}
and use the tool to install Python.
If Git is not available, this function will instead download the tool
directly from the repository and make it available to \code{reticulate}.
}
