% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment2.R
\docType{methods}
\name{show,simLists-method}
\alias{show,simLists-method}
\title{Show method for \code{simLists}}
\usage{
\S4method{show}{simLists}(object)
}
\arguments{
\item{object}{\code{simLists}}
}
\description{
Show method for \code{simLists}
}
\author{
Eliot McIntire
}
