% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-simInit.R
\name{simInitAndSpades}
\alias{simInitAndSpades}
\alias{simInitAndExperiment}
\title{Call \code{simInit} and \code{spades} together}
\usage{
simInitAndSpades(...)

simInitAndExperiment(...)
}
\arguments{
\item{...}{Arguments passed to simInit, and spades or experiment}
}
\value{
Same as \code{\link{spades}} (a \code{simList}) or
    \code{\link{experiment}} (list of \code{simList} objects)
}
\description{
These functions are convenience wrappers that may allow for
more efficient Caching.
Passes all arguments to \code{simInit}, then passes the created \code{simList}
to \code{spades} or \code{experiment}.
}
