% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simList-accessors.R
\name{newObjectsCreated}
\alias{newObjectsCreated}
\alias{objectsCreated}
\title{Show which objects were first created in a \code{simInit} or \code{spades} call}
\usage{
newObjectsCreated(sim)
}
\arguments{
\item{sim}{A \code{simList} object that contains data.table \code{objects}}
}
\value{
The \code{data.table} of the objects created, alongside the \code{current(sim)}
at each moment of creation.
}
\description{
This does an \code{rbindlist(sim$._objectsCreated)}.
This object in the \code{sim} records the yellow message that reports on when objects are created.
}
