% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spades-tools-deprecated.R
\name{smallNamify}
\alias{smallNamify}
\title{\code{SpaDES.tools} deprecated functions}
\usage{
smallNamify(name)
}
\arguments{
\item{name}{A character string giving the name/path of a file.}
}
\description{
\code{SpaDES.tools} deprecated functions
}
