% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat.R
\name{SpaDEStestMuffleOutput}
\alias{SpaDEStestMuffleOutput}
\title{SpaDES test: Muffle output}
\usage{
SpaDEStestMuffleOutput(
  expr,
  ...,
  suppressOutput = testthat::is_testing(),
  handleConditions = testthat::is_testing(),
  suppressWarnings = getOption("spades.test.suppressWarnings", default = FALSE)
)
}
\arguments{
\item{expr}{expression to be evaluated inside \code{\link{withCallingHandlers}}.}

\item{...}{optional additional arguments to \code{\link{withCallingHandlers}}.}

\item{suppressOutput}{logical. Sink output to a temporary file.}

\item{handleConditions}{logical. If FALSE, the expression will be evaluated
as normal outside of \code{\link{withCallingHandlers}}.
This is the default for interactive testing.}

\item{suppressWarnings}{logical. Suppress all warnings.
Default is FALSE or \code{getOption("spades.test.suppressWarnings")}}
}
\description{
A wrapper of \code{\link{withCallingHandlers}}
that intends to handle output, messages, and innocuous warnings from calls to
\code{\link[SpaDES.core]{simInit}},
\code{\link[SpaDES.core]{spades}},
and \code{\link[SpaDES.project]{setupProject}}.
All warnings can be silenced by setting \code{suppressWarnings = TRUE}
or \code{option("spades.test.suppressWarnings" = TRUE)}.
}
