% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat.R
\name{SpaDEStestSetGlobalOptions}
\alias{SpaDEStestSetGlobalOptions}
\title{SpaDES test: Set up global options}
\usage{
SpaDEStestSetGlobalOptions(
  reproducible.verbose = if (testthat::is_testing()) -2,
  Require.verbose = if (testthat::is_testing()) -2,
  Require.cloneFrom = Sys.getenv("R_LIBS_USER"),
  spades.moduleCodeChecks = if (testthat::is_testing()) FALSE,
  spades.moduleDocument = FALSE,
  SpaDES.project.updateRprofile = FALSE,
  teardownEnv = if (testthat::is_testing()) testthat::teardown_env()
)
}
\arguments{
\item{reproducible.verbose}{reproducible.verbose R package global option.}

\item{Require.verbose}{Require.verbose R package global option.}

\item{Require.cloneFrom}{Require.cloneFrom R package global option.}

\item{spades.moduleCodeChecks}{spades.moduleCodeChecks R package global option.}

\item{spades.moduleDocument}{spades.moduleDocument R package global option.}

\item{SpaDES.project.updateRprofile}{SpaDES.project R package global option.}

\item{teardownEnv}{environment. Optional. Environment to use for scoping.
The default for testing is the \code{testthat::teardown_env()}.}
}
\description{
Set local global options for testing. See \strong{Details}.
}
\details{
This function was designed to be called in a \code{tests/testthat/setup.R} file.

The defaults aim to provide an optimal environment for both standard
and interactive testing.
}
