% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat.R
\name{SpaDEStestSetUpEnvironment}
\alias{SpaDEStestSetUpEnvironment}
\title{SpaDES test: Set up environment}
\usage{
SpaDEStestSetUpEnvironment(
  localOutputSink = testthat::is_testing(),
  localMessageSink = FALSE,
  reproducible.verbose = if (testthat::is_testing()) -2,
  Require.verbose = if (testthat::is_testing()) -2,
  Require.cloneFrom = Sys.getenv("R_LIBS_USER"),
  spades.moduleCodeChecks = if (testthat::is_testing()) FALSE,
  spades.moduleDocument = FALSE,
  SpaDES.project.updateRprofile = FALSE,
  spadesTestPaths = .test_directories(),
  teardownEnv = if (testthat::is_testing()) testthat::teardown_env() else parent.frame(2)
)
}
\arguments{
\item{localOutputSink}{logical. Sink output to a temporary file}

\item{localMessageSink}{logical. Sink messages to a temporary file}

\item{reproducible.verbose}{reproducible.verbose R package global option.}

\item{Require.verbose}{Require.verbose R package global option.}

\item{Require.cloneFrom}{Require.cloneFrom R package global option.}

\item{spades.moduleCodeChecks}{spades.moduleCodeChecks R package global option.}

\item{spades.moduleDocument}{spades.moduleDocument R package global option.}

\item{SpaDES.project.updateRprofile}{SpaDES.project R package global option.}

\item{spadesTestPaths}{list. Optional.
List of test paths where \code{spadesTestPaths$temp$root} is the location
of the output and message sink files.}

\item{teardownEnv}{environment. Optional. Environment to set up.
The default for testing is the \code{testthat::teardown_env()}.}
}
\description{
Set local global options and sink output to file. See \strong{Details}.
}
\details{
This function was designed to be called in a \code{tests/testthat/setup.R} file
After \code{\link{SpaDEStestSetUpDirectories}}.
The defaults aim to provide an optimal environment for both standard and
interactive testing.
}
