% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simList-accessors.R
\name{startEnd}
\alias{startEnd}
\title{Simulation start and end times}
\usage{
startEnd(sim)
}
\arguments{
\item{sim}{A \code{simList} object.}
}
\value{
A named numeric of length 2, with elements \code{start} and \code{end}.
}
\description{
A convenience accessor returning the simulation \code{start} and \code{end} times as a
single named numeric vector of length 2, i.e.
\code{unlist(times(sim)[c("start", "end")])}.
}
\author{
Eliot McIntire
}
