% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.tryCatch}
\alias{.tryCatch}
\title{\code{tryCatch} that keeps warnings, errors and value (result)}
\usage{
.tryCatch(expr)
}
\description{
From \url{https://stackoverflow.com/a/24569739/3890027}
}
\keyword{internal}
