% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-parseModule.R
\name{.verb}
\alias{.verb}
\title{Chose verb conjugation for "to be"}
\usage{
.verb(item)
}
\arguments{
\item{item}{The item to accord conjugation with. If length 1, then "is"
else "are"}
}
\value{
"is" or "are"
}
\description{
Chose verb conjugation for "to be"
}
\keyword{internal}
