% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugging.R
\name{writeEventInfo}
\alias{writeEventInfo}
\title{Write simulation event info to file}
\usage{
writeEventInfo(sim, file = "events.txt", append = FALSE)
}
\arguments{
\item{sim}{A \code{simList} object.}

\item{file}{Character specifying the filename (default \file{"events.txt"}).}

\item{append}{Logical indicating whether to append to the file (default \code{FALSE}).}
}
\value{
Nothing returned. Invoked for its side-effect of writing to file.
}
\description{
Useful for debugging.
}
\author{
Alex Chubaty
}
