% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadInputs.R
\name{writeInputsOnDisk}
\alias{writeInputsOnDisk}
\title{Write module inputs on disk}
\usage{
writeInputsOnDisk(x, filename, rasterDatatype = NULL)
}
\arguments{
\item{x}{A \code{Spatial*}, \code{sf} or \code{Raster*} object.}

\item{filename}{Output filename}

\item{rasterDatatype}{Output data type. Passed to \code{\link[raster]{writeRaster}}}
}
\description{
Can be used to write prepared inputs on disk.
}
\author{
Eliot McIntire

Jean Marchal
}
