% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simList-accessors.R
\name{.addDepends}
\alias{.addDepends}
\alias{.addDepends,simList,.moduleDeps-method}
\title{Add simulation dependencies}
\usage{
.addDepends(sim, x)

\S4method{.addDepends}{simList,.moduleDeps}(sim, x)
}
\arguments{
\item{sim}{A \code{simList} object from which to extract element(s) or in which to replace element(s).}

\item{x}{A named list containing the parameters used to construct a new
\href{moduleDeps-class}{.moduleDeps} object.}
}
\value{
A \code{simList} object.
}
\description{
Internal function.
Adds a \href{moduleDeps-class}{.moduleDeps} object to the simulation dependency list.
}
\seealso{
Other functions to access elements of a \code{simList} object: 
\code{\link{checkpointFile}()},
\code{\link{envir}()},
\code{\link{events}()},
\code{\link{globals}()},
\code{\link{inputs}()},
\code{\link{modules}()},
\code{\link{objs}()},
\code{\link{packages}()},
\code{\link{params}()},
\code{\link{paths}()},
\code{\link{progressInterval}()},
\code{\link{times}()}
}
\author{
Alex Chubaty
}
\concept{functions to access elements of a \code{simList} object}
\keyword{internal}
