% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment2.R
\name{as.data.table.simLists}
\alias{as.data.table.simLists}
\title{Coerce elements of a simLists object to a data.table}
\usage{
\method{as.data.table}{simLists}(x, byRep = TRUE, vals,
  objectsFromSim = NULL, objectsFromOutputs = NULL, ...)
}
\arguments{
\item{byRep}{Should the data.table have a column labelled "rep", indicating replicate
number/label. Currently, only \code{TRUE} is accepted.}

\item{vals}{A character vector or list of object names to extract from each
simList, or a list of quoted expressions to calculate for each \code{simList},
or a mix of character and quoted expressions.}

\item{objectsFromSim}{Character vector of objects to extract from the simLists. If
omitted, it will extract all objects from each simList in order to calculate the
\code{vals}. This may have a computational cost.}

\item{objectsFromOutputs}{Character vector of objects to load from the
\code{outputs(sim)} prior to evaluating \code{vals}. If there already is an object
with that same name in the \code{simList}, then it will be overwritten with
the object loaded from \code{outputs(sim)}. If there are many objects with the
same name, specifically from several \code{saveTime} values in the \code{outputs(sim)},
these will all be loaded, one at a time, \code{vals} evaluated one at a time, and
all the unique values will be returned. A column, \code{saveTime}, will be
part of the returned value.}

\item{...}{Currently unused.}
}
\description{
This is particularly useful to build plots using the tidyverse,
e.g., \code{ggplot2}
}
