% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-methods.R
\name{bindrows}
\alias{bindrows}
\title{Simple wrapper around \code{data.table::rbindlist}}
\usage{
bindrows(...)
}
\arguments{
\item{...}{one or more \code{data.frame}, \code{data.table}, or \code{list} objects}
}
\value{
a \code{data.table} object
}
\description{
This simply sets defaults to \code{fill = TRUE}, and \code{use.names = TRUE}.
}
