% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.cacheMessage,simList-method}
\alias{.cacheMessage,simList-method}
\title{\code{.cacheMessage} for \code{simList} objects}
\usage{
\S4method{.cacheMessage}{simList}(
  object,
  functionName,
  fromMemoise = getOption("reproducible.useMemoise", TRUE)
)
}
\arguments{
\item{object}{Any R object.}

\item{functionName}{A character string indicating the function name}

\item{fromMemoise}{Logical. If \code{TRUE}, the message will be about
recovery from memoised copy}
}
\description{
See \link[reproducible:cacheMessage]{reproducible::.cacheMessage}.
}
\seealso{
\link[reproducible:cacheMessage]{reproducible::.cacheMessage}
}
