% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.checkCacheRepo,list-method}
\alias{.checkCacheRepo,list-method}
\title{\code{.checkCacheRepo} for \code{simList} objects}
\usage{
\S4method{.checkCacheRepo}{list}(object, create = FALSE)
}
\value{
character string representing a directory path to the cache repo
}
\description{
See \code{\link[reproducible:.checkCacheRepo]{reproducible::.checkCacheRepo()}}.
}
\seealso{
\link[reproducible:.checkCacheRepo]{reproducible::.checkCacheRepo}
}
