% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.checkCacheRepo,list-method}
\alias{.checkCacheRepo,list-method}
\title{\code{.checkCacheRepo} for \code{simList} objects}
\usage{
\S4method{.checkCacheRepo}{list}(object, create = FALSE)
}
\arguments{
\item{object}{Any R object returned from a function}

\item{create}{Logical. If TRUE, then it will create the path for cache.}
}
\value{
character string representing a directory path to the cache repo
}
\description{
See \link[reproducible:exportedMethods]{reproducible::.checkCacheRepo}.
}
\seealso{
\link[reproducible:exportedMethods]{reproducible::.checkCacheRepo}
}
