% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.checkCacheRepo,list-method}
\alias{.checkCacheRepo,list-method}
\title{\code{.checkCacheRepo} for \code{simList} objects}
\usage{
\S4method{.checkCacheRepo}{list}(object, create = FALSE)
}
\arguments{
\item{object}{An R object}

\item{create}{Logical. If TRUE, then it will create the path for cache.}
}
\description{
See \code{\link[reproducible]{.checkCacheRepo}}.
}
\seealso{
\code{\link[reproducible]{.checkCacheRepo}}
}
