% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-template.R
\name{checkModulePath}
\alias{checkModulePath}
\title{Check module path}
\usage{
checkModulePath()
}
\description{
Will compare default in \code{spadesOptions()} (i.e., \code{spades.modulePath}) to \code{getPaths()}.
Uses \code{"."} if \code{modulePath} not set (e.g., via \code{setPaths}).
If that is the case, then use \code{getPaths()[["modulePath"]]}.
}
