% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{clearCache,simList-method}
\alias{clearCache,simList-method}
\alias{showCache,simList-method}
\alias{keepCache,simList-method}
\title{\code{clearCache} for \code{simList} objects}
\usage{
\S4method{clearCache}{simList}(
  x,
  userTags = character(),
  after = NULL,
  before = NULL,
  fun = NULL,
  cacheId = NULL,
  ask = getOption("reproducible.ask"),
  useCloud = FALSE,
  cloudFolderID = getOption("reproducible.cloudFolderID", NULL),
  drv = getDrv(getOption("reproducible.drv", NULL)),
  conn = getOption("reproducible.conn", NULL),
  verbose = getOption("reproducible.verbose"),
  ...
)

\S4method{showCache}{simList}(
  x,
  userTags = character(),
  after = NULL,
  before = NULL,
  fun = NULL,
  cacheId = NULL,
  drv = getDrv(getOption("reproducible.drv", NULL)),
  conn = getOption("reproducible.conn", NULL),
  verbose = getOption("reproducible.verbose"),
  ...
)

\S4method{keepCache}{simList}(
  x,
  userTags = character(),
  after = NULL,
  before = NULL,
  ask = getOption("reproducible.ask"),
  drv = getDrv(getOption("reproducible.drv", NULL)),
  conn = getOption("reproducible.conn", NULL),
  verbose = getOption("reproducible.verbose"),
  ...
)
}
\arguments{
\item{drv}{an object that inherits from \code{DBIDriver}, or an existing
\code{DBIConnection} object (in order to clone an existing connection).}

\item{conn}{A \code{DBIConnection} object, as returned by \code{dbConnect()}.}
}
\value{
A \code{data.table} object showing the subset of items in the cache, located at \code{cachePath}
of the \code{sim} object, if \code{sim} is provided, or located in \code{cachePath}.
For \code{clearCache} (invoked for its side effect of clearing objects matching \code{userTags}, or those
between \code{after} or \code{before}), the returned \code{data.table} shows the removed items (invisibly).
}
\description{
This will take the \code{cachePath(object)} and pass
}
