% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{clearCacheEventsOnly}
\alias{clearCacheEventsOnly}
\title{Convenience wrapper around \code{clearCache} for SpaDES events}
\usage{
clearCacheEventsOnly(
  ask,
  x = getOption("reproducible.cachePath"),
  verbose = getOption("reproducible.verbose")
)
}
\value{
A list of individual \code{clearCache} outputs, one for each event that was
cleared.
}
\description{
This will clear only the event- and module-level caching that is triggered
using a module parameter, \code{.useCache}.
}
