% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadInputs.R
\name{cropReprojInputs}
\alias{cropReprojInputs}
\title{Reproject, crop a \code{Spatial*} or \code{Raster*} object}
\usage{
cropReprojInputs(x, studyArea = NULL, rasterToMatch = NULL, ...)
}
\arguments{
\item{x}{A \code{Spatial*}, \code{sf}, or \code{Raster*} object.}

\item{studyArea}{Template \code{SpatialPolygons*} object used for masking, after cropping.
If not in same CRS, then it will be \code{spTransform}ed to
CRS of \code{x} before masking. Currently, this function will not reproject the
\code{x}. \code{\link{postProcess.spatialObjects}}}

\item{rasterToMatch}{Template \code{Raster*} object used for cropping (so extent should be
the extent of desired outcome), reprojecting (including changing the
resolution and projection). See details in
\code{\link{postProcess.spatialObjects}}.}

\item{...}{Passed to \code{projectRaster} and \code{Cache}
cropping.}
}
\description{
This function can be used to crop or reproject module inputs from raw data.
}
\author{
Eliot McIntire

Jean Marchal
}
