% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.dealWithClass.simList}
\alias{.dealWithClass.simList}
\alias{.dealWithClassOnRecovery.simList}
\title{Methods for .dealWithClass and .dealWithClassOnRecovery}
\usage{
.dealWithClass.simList(
  obj,
  cachePath,
  drv = getOption("reproducible.drv", RSQLite::SQLite()),
  conn = getOption("reproducible.conn", NULL),
  verbose = getOption("reproducible.verbose")
)

.dealWithClassOnRecovery.simList(
  obj,
  cachePath,
  cacheId,
  drv = getOption("reproducible.drv", RSQLite::SQLite()),
  conn = getOption("reproducible.conn", NULL)
)
}
\value{
The same obj as passed into the function, but dealt with so that it can be
saved to disk.
}
\description{
Methods for .dealWithClass and .dealWithClassOnRecovery
}
