% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.wrap.simList}
\alias{.wrap.simList}
\alias{.wrap..simDeps}
\alias{.unwrap..simDeps}
\alias{.wrap..moduleDeps}
\alias{.unwrap..moduleDeps}
\alias{.unwrap.simList}
\title{Methods for \code{.wrap} and \code{.unwrap}}
\usage{
\method{.wrap}{simList}(
  obj,
  cachePath,
  preDigest,
  drv = getOption("reproducible.drv", NULL),
  conn = getOption("reproducible.conn", NULL),
  verbose = getOption("reproducible.verbose"),
  outputObjects = NULL,
  cacheId,
  ...
)

\method{.wrap}{.simDeps}(obj, ...)

\method{.unwrap}{.simDeps}(obj, ...)

\method{.wrap}{.moduleDeps}(obj, ...)

\method{.unwrap}{.moduleDeps}(obj, ...)

\method{.unwrap}{simList}(
  obj,
  cachePath,
  cacheId,
  drv = getOption("reproducible.drv", NULL),
  conn = getOption("reproducible.conn", NULL),
  ...
)
}
\arguments{
\item{drv}{an object that inherits from \code{DBIDriver}, or an existing
\code{DBIConnection} object (in order to clone an existing connection).}

\item{conn}{A \code{DBIConnection} object, as returned by \code{dbConnect()}.}
}
\value{
The same object as passed into the function, but dealt with so that it can be
saved to disk.
}
\description{
Methods for \code{.wrap} and \code{.unwrap}
}
