% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spades-core-deprecated.R
\name{experiment}
\alias{experiment}
\alias{experiment2}
\alias{POM}
\alias{simInitAndExperiment}
\title{Deprecated functions}
\usage{
experiment(...)

experiment2(...)

POM(...)

simInitAndExperiment(...)
}
\arguments{
\item{...}{Unused.}
}
\description{
These functions have been moved to \code{SpaDES.experiment} package.
}
