% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-repository.R
\docType{methods}
\name{downloadData}
\alias{downloadData}
\alias{downloadData,character,character,logical-method}
\alias{downloadData,character,missing,missing-method}
\alias{downloadData,character,missing,logical-method}
\alias{downloadData,character,character,missing-method}
\title{Download module data}
\usage{
downloadData(module, path, quiet, quickCheck = FALSE)

\S4method{downloadData}{character,character,logical}(module, path, quiet,
  quickCheck = FALSE)

\S4method{downloadData}{character,missing,missing}(module, quickCheck)

\S4method{downloadData}{character,missing,logical}(module, quiet, quickCheck)

\S4method{downloadData}{character,character,missing}(module, path, quickCheck)
}
\arguments{
\item{module}{Character string giving the name of the module.}

\item{path}{Character string giving the path to the module directory.}

\item{quiet}{Logical. This is passed to \code{download.file}. Default is FALSE.}

\item{quickCheck}{Logical. If \code{TRUE}, then the check with local data will only
use \code{file.size} instead of \code{digest::digest}. This is
faster, but less robust.}
}
\value{
Invisibly, a list of downloaded files.
}
\description{
Download external data for a module if not already present in the module
directory, or if there is a checksum mismatch indicating that the file is not
the correct one.
}
\details{
\code{downloadData} requires a checksums file to work, as it will only download
 the files specified therein. Hence, module developers should make sure they
 have manually downloaded all the necessary data and ran \code{checksums} to
 build a checksums file.
}
\seealso{
\code{\link{checksums}} for building a checksums file.
}
\author{
Alex Chubaty
}
