% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadInputs.R
\name{downloadFromWebDB}
\alias{downloadFromWebDB}
\title{Download file from web databases}
\usage{
downloadFromWebDB(filename, filepath, dataset = NULL)
}
\arguments{
\item{filename}{Character string naming the file to be downloaded.}

\item{filepath}{Character string giving the path where the file will be
written.}

\item{dataset}{Optional character string representing the dataset of interest
for download. Allows for restricting the lookup for the url to a dataset,
thus avoiding filename collision.}
}
\description{
This function can be used to download a file from a web database listed in
\code{webDatabases:::urls}.
}
\author{
Jean Marchal
}
