% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment2.R
\docType{methods}
\name{experiment2}
\alias{experiment2}
\alias{experiment2,simList-method}
\title{Run experiment, algorithm 2, using \code{\link{spades}}}
\usage{
experiment2(..., replicates = 1, clearSimEnv = FALSE,
  createUniquePaths = c("outputPath"), useCache = FALSE)

\S4method{experiment2}{simList}(..., replicates = 1,
  clearSimEnv = FALSE, createUniquePaths = c("outputPath"),
  useCache = FALSE)
}
\arguments{
\item{...}{One or more \code{simList} objects}

\item{replicates}{The number of replicates to run of the same \code{simList}.
See details and examples.}

\item{clearSimEnv}{Logical. If TRUE, then the envir(sim) of each simList in the return list
is emptied. This is to reduce RAM load of large return object.
Default FALSE.}

\item{createUniquePaths}{A character vector of the \code{paths} passed to \code{simInit},
indicating which should create a new, unique path, as a sub-path to the original
\code{paths} of \code{simInit}. Default, and currently only option, is \code{"outputPath"}}

\item{useCache}{Logical. Passed to \code{spades}. This will be passed with the \code{simList}
name and replicate number, allowing each replicate and each \code{simList} to be
seen as a non-cached call to \code{spades}. This will, however, may prevent \code{spades}
calls from running a second time during second call to the same
\code{experiment2} function.}
}
\value{
Invisibly returns a \code{simLists} object. This class
extends the \code{environment} class and
contains \code{simList} objects.
}
\description{
Given one or more \code{simList} objects, run a series of \code{spades} calls
in a structured, organized way. Methods are available to deal with outputs,
such as \code{as.data.table.simList} which can pull out simple to complex
values from every resulting \code{simList} or object saved by \code{outputs}
in every \code{simList} run. This uses \code{future} internally, allowing
for various backends and parallelism.0
}
\details{
The \code{outputPath} is changed so that every simulation puts outputs in a sub-directory
of the original \code{outputPath} of each \code{simList}.
}
\seealso{
\code{\link{simInit}}, \code{\link{spades}}, \code{\link{experiment}}
}
\author{
Eliot McIntire
}
