% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadData.R
\docType{methods}
\name{extractURL}
\alias{extractURL}
\alias{extractURL,character,missing-method}
\alias{extractURL,character,simList-method}
\title{Extract a url from module metadata}
\usage{
extractURL(objectName, sim, module)

\S4method{extractURL}{character,missing}(objectName, sim, module)

\S4method{extractURL}{character,simList}(objectName, sim, module)
}
\arguments{
\item{objectName}{A character string of the object name in the metadata.}

\item{sim}{A \code{simList} object from which to extract the \code{sourceURL}}

\item{module}{An optional character string of the module name whose metadata is
to be used. If omitted, the function will use the \code{currentModule(sim)},
if defined.}
}
\value{
The url.
}
\description{
This will get the \code{sourceURL} for the object named.
}
\author{
Eliot McIntire
}
