% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadData.R
\docType{methods}
\name{extractURL}
\alias{extractURL}
\alias{extractURL,character,simList-method}
\title{Extract a url from module metadata}
\usage{
extractURL(objectName, sim, module)

\S4method{extractURL}{character,simList}(objectName, sim, module)
}
\arguments{
\item{objectName}{A character string of the object name in the metadata.}

\item{sim}{A \code{simList} object from which to extract the \code{sourceURL}.}

\item{module}{Character giving the name of the module.}
}
\value{
The url.
}
\description{
This will get the sourceURL for the object named.
}
\author{
Eliot McIntire
}
