% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-simInit.R
\name{.findModuleName}
\alias{.findModuleName}
\title{Identify module names up to a given recursive level}
\usage{
.findModuleName(modList, recursive = 0)
}
\arguments{
\item{recursive}{Numeric. The depth of recursion, where 0 is only top level, 1 is 1 level in etc.}

\item{modules}{(Nested) Named list of module names}
}
\value{
Character vector of modules names
}
\description{
With children, parents, grandparents, etc, there can be severl "layers" of recursion.
Some functions need to evaluate the outer level for a value, if found, they don't need
to proceed further. If not found, increment one more level of recursion, etc.
}
\keyword{internal}
