% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-checking.R
\name{.findAllSims}
\alias{.findAllSims}
\alias{.findAssignSims}
\alias{.findGetSims}
\title{Find all references to sim$}
\usage{
.findAllSims(envToFindSim, moduleEnv, type)

.findAssignSims(x)

.findGetSims(x)
}
\arguments{
\item{envToFindSim}{An environment where sim is defined}

\item{moduleEnv}{The environment where the module functions are}

\item{type}{Either "get" or "assign" indicating which side of <- we are searching}

\item{x}{A call in which to search for sim}
}
\value{
A character string with all sim objects found
}
\description{
Find all references to sim$
}
\author{
Eliot McIntire
}
\keyword{internal}
