% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadInputs.R
\name{fixErrors.SpatialPolygons}
\alias{fixErrors.SpatialPolygons}
\title{Fix \code{rgeos::gIsValid} failures in \code{SpatialPolygons}}
\usage{
\method{fixErrors}{SpatialPolygons}(x, targetFile, attemptErrorFixes = TRUE,
  ...)
}
\description{
This uses \code{raster::buffer(..., width = 0)} internally, which fixes some
failures to \code{rgeos::gIsValid}
}
