% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-repository.R
\docType{methods}
\name{getModuleVersion}
\alias{getModuleVersion}
\alias{getModuleVersion,character,character-method}
\alias{getModuleVersion,character,missing-method}
\title{Find the latest module version from a SpaDES module repository}
\usage{
getModuleVersion(name, repo)

\S4method{getModuleVersion}{character,character}(name, repo)

\S4method{getModuleVersion}{character,missing}(name)
}
\arguments{
\item{name}{Character string giving the module name.}

\item{repo}{GitHub repository name, specified as \code{"username/repositoryname"}.
Default is \code{"PredictiveEcology/SpaDES-modules"}, which is
specified by the global option \code{spades.moduleRepo}.
Only \code{master} branches can be used at this point.}
}
\description{
Modified from \url{http://stackoverflow.com/a/25485782/1380598}.
}
\details{
\code{getModuleVersion} extracts a module's version from the module .zip
         file name.
         Module .zip files are searched inside their respective module
         folders, which should themselves be in a \code{"/modules"} folder in a
         GitHub public repository
}
\seealso{
\code{\link{zipModule}} for creating module .zip folders.
}
\author{
Alex Chubaty
}
