% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simList-accessors.R
\name{globals}
\alias{globals}
\alias{globals,simList-method}
\alias{globals<-}
\alias{globals<-,simList-method}
\alias{G}
\alias{G,simList-method}
\alias{G<-}
\alias{G<-,simList-method}
\title{Get and set global simulation parameters}
\usage{
globals(sim)

\S4method{globals}{simList}(sim)

globals(sim) <- value

\S4method{globals}{simList}(sim) <- value

G(sim)

\S4method{G}{simList}(sim)

G(sim) <- value

\S4method{G}{simList}(sim) <- value
}
\arguments{
\item{sim}{A \code{simList} object from which to extract element(s) or in which to replace element(s).}

\item{value}{The parameter value to be set (in the corresponding \code{module} and \code{param}).}
}
\value{
For \code{globals()} and \code{G()}, a named list of the global parameters, i.e., the
\code{.globals} element of the \code{params} slot. For the replacement methods
\verb{globals<-} and \verb{G<-}, the updated \code{simList}, with any module parameters that
take their value from a global re-synchronised.
}
\description{
\code{globals}, and the alias \code{G}, accesses or sets the "globals"
in the \code{simList}. This currently is not an explicit slot in the \code{simList},
but it is a \code{.globals} element in the \code{params} slot of the \code{simList}.
}
\seealso{
\link{SpaDES.core-package}, specifically the section 1.2.1 on Simulation Parameters.

Other functions to access elements of a 'simList' object:
\code{\link[=.addDepends]{.addDepends()}},
\code{\link[=checkpointFile]{checkpointFile()}},
\code{\link[=envir]{envir()}},
\code{\link[=events]{events()}},
\code{\link[=inputs]{inputs()}},
\code{\link[=modules]{modules()}},
\code{\link[=objs]{objs()}},
\code{\link[=packages]{packages()}},
\code{\link[=params]{params()}},
\code{\link[=paths]{paths()}},
\code{\link[=progressInterval]{progressInterval()}},
\code{\link[=times]{times()}}
}
\concept{functions to access elements of a 'simList' object}
