% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadSimList}
\alias{loadSimList}
\title{Load a saved \code{simList} from file}
\usage{
loadSimList(file)
}
\arguments{
\item{file}{Character giving the name of a saved simulation file}
}
\description{
Load a saved \code{simList} from file
}
