% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{makeMemoisable.simList}
\alias{makeMemoisable.simList}
\alias{unmakeMemoisable.simList_}
\title{Make \code{simList} correctly work with \code{memoise}}
\usage{
\method{makeMemoisable}{simList}(x)

\method{unmakeMemoisable}{simList_}(x)
}
\arguments{
\item{x}{An object to make memoisable.
See individual methods in other packages.}
}
\value{
A \code{simList_} object or a \code{simList}, in the case of \code{unmakeMemoisable}.
}
\description{
Because of the environment slot, \code{simList} objects don't correctly
memoise a \code{simList}.
This method for \code{simList} converts the object to a \code{simList_} first.
}
\seealso{
\code{\link[reproducible:exportedMethods]{reproducible::makeMemoisable()}}
}
