% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{memoryUse}
\alias{memoryUse}
\title{Show memory use}
\usage{
memoryUse(sim, max = TRUE)
}
\arguments{
\item{sim}{A completed simList}

\item{max}{Logical. If TRUE, then it the return value will be summarized by
module/event, showing the maximum memory used. If \code{FALSE}, then
the raw memory used during each event will be shown.}
}
\description{
This will only work if the user has specified before running
the \code{spades} call, set the interval, in seconds, that ps is
run with \code{options("spades.memoryUseInterval" = 1)}. The default
is 0, meaning no interval, "off".
}
