% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-define.R
\docType{data}
\name{moduleDefaults}
\alias{moduleDefaults}
\title{Defaults values used in defineModule}
\format{
An object of class \code{list} of length 12.
}
\usage{
moduleDefaults
}
\description{
Where individual elements are missing in \code{defineModule},
these defaults will be used.
}
\keyword{datasets}
