% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-template.R
\name{newProjectCode}
\alias{newProjectCode}
\title{Create new module code file}
\usage{
newProjectCode(name, path, open)
}
\arguments{
\item{name}{Character string specifying the name of the new module.}

\item{path}{Character string. Subdirectory in which to place the new module code file.
The default is the current working directory.}

\item{open}{Logical. Should the new project file be opened after creation?
Default \code{TRUE} in an interactive session.}
}
\description{
Create new module code file
}
\author{
Alex Chubaty
}
