% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-template.R
\docType{methods}
\name{newProjectCode}
\alias{newProjectCode}
\alias{newProjectCode,character,character,logical-method}
\title{Create new module code file}
\usage{
newProjectCode(name, path, open)

\S4method{newProjectCode}{character,character,logical}(name, path,
  open = interactive())
}
\arguments{
\item{name}{project name (name of project directory)}

\item{path}{path to directory in which to create the project directory}

\item{open}{Logical. Should the new project file be opened after creation?
Default \code{TRUE} in an interactive session.}
}
\description{
Create new module code file
}
\author{
Alex Chubaty
}
