% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{objectSize.simList}
\alias{objectSize.simList}
\title{Object size for simLists}
\usage{
\method{objectSize}{simList}(x)
}
\arguments{
\item{x}{An object}
}
\description{
Recursively, runs \code{object\.size} on the simList environment.
Currently, this will not assess object.size of the other elements
}
\examples{
a <- simInit(objects = list(d = 1:10, b = 2:20))
objectSize(a)
object.size(a)
}
