% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{objSize.simList}
\alias{objSize.simList}
\title{Object size for simLists}
\usage{
\method{objSize}{simList}(x, quick = getOption("reproducible.quick", FALSE))
}
\arguments{
\item{x}{An object}

\item{quick}{Logical. Only some methods use this. e.g.,
\code{Path} class objects. In which case, \code{file.size} will be
used instead of \code{object.size}.}
}
\description{
Recursively, runs \code{object\.size} on the simList environment.
Currently, this will not assess object.size of the other elements
}
\examples{
a <- simInit(objects = list(d = 1:10, b = 2:20))
objSize(a)
object.size(a)
}
