% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{.parseElems,simList-method}
\alias{.parseElems,simList-method}
\title{\code{.parseElems} for \code{simList} class objects}
\usage{
\S4method{.parseElems}{simList}(tmp, elems, envir)
}
\arguments{
\item{tmp}{A evaluated object}

\item{elems}{A character string to be parsed}

\item{envir}{An environment}
}
\value{
An object, parsed from a character string and an environment.
}
\description{
See \link[quickPlot:parseElems]{quickPlot::.parseElems}.
}
\seealso{
\link[quickPlot:parseElems]{quickPlot::.parseElems}
}
