% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\docType{data}
\name{.pkgEnv}
\alias{.pkgEnv}
\title{The \code{SpaDES.core} package environment}
\format{An object of class \code{environment} of length 7.}
\usage{
.pkgEnv
}
\description{
Environment used internally to store internal package objects and methods.
}
\keyword{internal}
