% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.preDigestByClass,simList-method}
\alias{.preDigestByClass,simList-method}
\title{Pre-digesting method for \code{simList}}
\usage{
\S4method{.preDigestByClass}{simList}(object)
}
\arguments{
\item{object}{Any R object.}
}
\description{
Takes a snapshot of \code{simList} objects.
}
\details{
See \code{\link[reproducible:preDigestByClass]{.preDigestByClass}}.
}
\seealso{
\code{\link[reproducible:preDigestByClass]{.preDigestByClass}}
}
\author{
Eliot McIntire
}
