% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.prepareOutput,simList-method}
\alias{.prepareOutput,simList-method}
\title{\code{.prepareOutput} for \code{simList} objects}
\usage{
\S4method{.prepareOutput}{simList}(object, cacheRepo, ...)
}
\arguments{
\item{object}{Any R object returned from a function}

\item{...}{Anything passed to methods.}
}
\description{
See \link[reproducible:exportedMethods]{reproducible::.prepareOutput}.
}
\seealso{
\link[reproducible:exportedMethods]{reproducible::.prepareOutput}
}
