% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.prepareOutput,simList-method}
\alias{.prepareOutput,simList-method}
\title{\code{.prepareOutput} for \code{simList} objects}
\usage{
\S4method{.prepareOutput}{simList}(object, cacheRepo, ...)
}
\arguments{
\item{object}{Any R object}

\item{cacheRepo}{A repository used for storing cached objects.
This is optional if \code{Cache} is used inside a SpaDES module.}

\item{...}{Arguments passed to \code{FUN}}
}
\description{
See \code{\link[reproducible:prepareOutput]{.prepareOutput}}.
}
\seealso{
\code{\link[reproducible:prepareOutput]{.prepareOutput}}
}
