% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadInputs.R
\name{prepInputs}
\alias{prepInputs}
\title{Download and optionally post process files}
\usage{
prepInputs(targetFile, url = NULL, archive = NULL, alsoExtract = NULL,
  destinationPath = ".", fun = "raster::raster",
  quick = getOption("reproducible.quick"), overwrite = FALSE,
  purge = FALSE, ...)
}
\arguments{
\item{targetFile}{Character string giving the path to the eventual
file (raster, shapefile, csv, etc.) after downloading and extracting from
a zip or tar archive. This is the file \emph{before}
it is passed to \code{postProcess}. Currently, the internal
checksumming does not checksum the file after it is
\code{postProcess}ed (e.g., cropped/reprojected/masked).
Using \code{Cache} around \code{prepInputs} will do a
sufficient job in these cases.}

\item{url}{Optional character string indicating the URL to download from. Normally,
if used within a module, this url should be explicitly given as sourceURL for an
\code{expectsInput}. In that case, it will use the module's checksums file to
confirm that the download occurred correctly. If URL is used here, an ad hoc
checksums will be created in the \code{destinationPath}. This will be used in
subsequent calls to \code{prepInputs}, comparing the file on hand with the ad hoc
\code{checksums.txt}.}

\item{archive}{Optional character string giving the path of an archive
containing \code{targetFile}, or a vector giving a set of nested archives
(e.g., \code{c("xxx.tar", "inner.zip")}). If there is/are (an) inner archive(s),
but they are unknown, the function will try all until it finds the
\code{targetFile}}

\item{alsoExtract}{Optional character string naming files other than
\code{targetFile} that must be extracted from the \code{archive}.}

\item{destinationPath}{Character string of where to download to, and do
all writing of files in.}

\item{fun}{Character string indicating the function to use to load \code{targetFile} into
an \code{R} object.}

\item{quick}{Logical. This is passed internally to \code{\link{checksums}} and
\code{\link{downloadData}} (the quickCheck argument for both),
and to \code{\link{Cache}} (the quick argument). This results in
faster, though less robust checking of inputs. See the respective
functions.}

\item{overwrite}{Logical. Should downloading and all the other actions occur even if they
pass the checksums or the files are all there.}

\item{purge}{When prepInputs is called from outside a module, it will write a \code{CHECKSUMS.txt}
file. If there is an incorrect \code{CHECKSUMS.txt}, this will purge it.}

\item{...}{Additional arguments that can be passed to
\code{fun}, \code{\link{fixErrors}} and \code{\link{postProcess}} and
also \code{\link[reproducible]{Cache}}
(if \code{useCache = TRUE}, which is the default unless
\code{options(reproducible.useCache = FALSE)}). Since \code{...} is passed to
\code{\link{postProcess}}, these may be passed into other functions.
See details and examples.}
}
\description{
This function can be used to prepare R objects from remote or local data sources.
The object of this function is to provide a reproducible version of a series of
commonly used steps for getting, loading, and processing data. In the case of
\code{Spatial*} and \code{Raster*} objects, then the postprocessing may include
cropping, reprojecting, masking, if the appropriate arguments are used. See examples.
}
\details{
\code{prepInputs} has some functionality of its own; however, its primary purpose
is to run in sequence: \code{downloadFile}, \code{extractFromArchive}, \code{fun},
\code{postProcess}.
runs several other functions, conditionally and sequentially:
\code{downloadFromWebDB} or \code{downloadData} if used within a module. If used
outside of a SpaDES module, then it will use \code{file.download} or
\code{googledrive::drive_download} if the \code{url} has \code{https://drive.google.com}.
If the download is a .zip or .tar file (i.e., an archive), then the function will
run \code{extractFromArchive}. Then it will sequentially try to load the extracted file (if
passed as \code{targetFile}). If the default \code{fun} is not left as is,
the function will try to use those. If the file is not a raster file, it will try
using raster::shapefile. NOTE: This function is still experimental: use
with caution.
}
\section{3. postProcessing of \code{Raster*} and \code{Spatial*} objects}{


If \code{rasterToMatch} or \code{studyArea} are used, then this will trigger several
subsequent functions, specifically the sequence, \emph{Crop, reproject, mask}, which appears
to be a common sequence in spatial simulation. See \code{\link{postProcess.spatialObjects}}.
}

\examples{
# This function works within a module, when "sourceURL" is supplied
#   in the metadata in the "expectsInputs(..., sourceURL = ""), but can
# also run outside a module, e.g., with url argument.
\dontrun{
# Put chunks like this in your .inputObjects
if (!suppliedElsewhere("test", sim))
  sim$test <- Cache(prepInputs, "raster.tif", "downloadedArchive.zip",
                    destinationPath = dataPath(sim), studyArea = sim$studyArea,
                    rasterToMatch = sim$otherRasterTemplate, overwrite = TRUE)

# download a zip file from internet, unzip all files, load as shapefile, Cache the call
# First time: don't know all files - prepInputs will guess, if download file is an archive,
#   then extract all files, then if there is a .shp, it will load with raster::shapefile
dPath <- file.path(tempdir(), "ecozones")
shpEcozone <- prepInputs(destinationPath = dPath,
                     url = "http://sis.agr.gc.ca/cansis/nsdb/ecostrat/zone/ecozone_shp.zip")

# Robust to partial file deletions:
unlink(dir(dPath, full.names = TRUE)[1:3])
shpEcozone <- prepInputs(destinationPath = dPath,
                     url = "http://sis.agr.gc.ca/cansis/nsdb/ecostrat/zone/ecozone_shp.zip")
unlink(dPath, recursive = TRUE)

# Once this is done, can be more precise in operational code:
#  specify targetFile, alsoExtract, and fun, wrap with Cache
ecozoneFilename <- file.path(dPath, "ecozones.shp")
ecozoneFiles <- c("ecozones.dbf", "ecozones.prj",
                  "ecozones.sbn", "ecozones.sbx", "ecozones.shp", "ecozones.shx")
shpEcozone <- prepInputs(targetFile = asPath(ecozoneFilename),
                    url = "http://sis.agr.gc.ca/cansis/nsdb/ecostrat/zone/ecozone_shp.zip",
                    alsoExtract = asPath(ecozoneFiles),
                    fun = "shapefile", destinationPath = dPath)
unlink(dPath, recursive = TRUE)

#' # Add a study area to Crop and Mask to
# Create a "study area"
library(SpaDES.tools)
StudyArea <- randomPolygon(x = sp::SpatialPoints(matrix(c(-110, 60), ncol=2)), 1e8)

#  specify targetFile, alsoExtract, and fun, wrap with Cache
ecozoneFilename <- file.path(dPath, "ecozones.shp")
# Note, you don't need to "alsoExtract" the archive... if the archive is not there, but the
#   targetFile is there, it will not redownload the archive.
ecozoneFiles <- c("ecozones.dbf", "ecozones.prj",
                  "ecozones.sbn", "ecozones.sbx", "ecozones.shp", "ecozones.shx")
shpEcozoneSm <- Cache(prepInputs,
                         url = "http://sis.agr.gc.ca/cansis/nsdb/ecostrat/zone/ecozone_shp.zip",
                         targetFile = asPath(ecozoneFilename),
                         alsoExtract = asPath(ecozoneFiles),
                         studyArea = StudyArea,
                         fun = "shapefile", destinationPath = dPath)

dev();
Plot(shpEcozone)
Plot(shpEcozoneSm, addTo = "shpEcozone", gp = gpar(col = "red"))
unlink(dPath)

# Big Raster, with crop and mask to Study Area - no reprojecting (lossy) of raster,
#   but the StudyArea does get reprojected, need to use rasterToMatch
lcc2005Filename <- file.path(dPath, "LCC2005_V1_4a.tif")
url <- file.path("ftp://ftp.ccrs.nrcan.gc.ca/ad/NLCCLandCover",
                 "LandcoverCanada2005_250m/LandCoverOfCanada2005_V1_4.zip")
dPath <- file.path(tempdir(), "LCC")

# messages received below may help for filling in more arguments in the subsequent call
LCC2005 <- Cache(prepInputs, url = url,
                     destinationPath = asPath(dPath),
                     studyArea = StudyArea)

Plot(LCC2005)

# Specifying more args
LCC2005 <- Cache(prepInputs, url = url,
                     targetFile = lcc2005Filename,
                     archive = asPath("LandCoverOfCanada2005_V1_4.zip"),
                     destinationPath = asPath(dPath),
                     studyArea = StudyArea)
}

}
\seealso{
\code{\link{downloadFile}}, \code{\link{extractFromArchive}},
         \code{\link{downloadFile}},  \code{\link{postProcess}}.
}
\author{
Eliot McIntire

Jean Marchal
}
