% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-repository.R
\name{remoteFileSize}
\alias{remoteFileSize}
\title{Determine the size of a remotely hosted file}
\usage{
remoteFileSize(url)
}
\arguments{
\item{url}{The url of the remote file.}
}
\value{
A numeric indicating the sive of the remote file in bytes.
}
\description{
Query a remote web server to determine the size of a remote file.
}
\examples{
urls <- c("https://www.alexchubaty.com/uploads/2011/11/open-forest-science-journal.csl",
          "https://www.alexchubaty.com/uploads/2011/08/models_GUI_2011-08-07.zip",
          "http://example.com/doesntexist.csv")
try(remoteFileSize(urls))

}
\author{
Eliot McIntire and Alex Chubaty
}
