% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{restartR}
\alias{restartR}
\title{Restart R programmatically}
\usage{
restartR(reloadPkgs = TRUE)
}
\arguments{
\item{reloadPkgs}{Logical. If \code{TRUE}, it will attempt to reload all the packages
as they were in previous session, in the same order. If \code{FALSE}, it will
load no packages beyond normal R startup. Default \code{TRUE}}
}
\description{
This will attempt to restart the R session, reloading all packages.
The main purpose for doing this is to clear memory leaks that are not
yet diagnosed.
}
