% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python-ReticulateFindPython.R
\name{reticulate_install_python_windows}
\alias{reticulate_install_python_windows}
\title{Reticulate: Install Python on Windows}
\usage{
reticulate_install_python_windows(
  version = NULL,
  pyenvRoot = NULL,
  useGit = TRUE,
  prompt = interactive()
)
}
\arguments{
\item{version}{character. Python version string.}

\item{pyenvRoot}{character. Path of location for install of the pyenv-win tool.
Defaults to the R user data directory.}

\item{useGit}{logical. Allow \code{reticulate} to clone pyenv-win if Git is available.}

\item{prompt}{logical. Prompt user to approve download of pyenv-win tool.}
}
\description{
Download and use the \href{https://pypi.org/project/pyenv-win/}{pyenv-win}
Python version management tool from Github if necessary.
}
\keyword{internal}
