% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.robustDigest,simList-method}
\alias{.robustDigest,simList-method}
\alias{Cache}
\title{\code{.robustDigest} for \code{simList} objects}
\usage{
\S4method{.robustDigest}{simList}(object, .objects, length, algo = "xxhash64", quick, classOptions)
}
\description{
This is intended to be used within the \code{Cache} function, but can be used to evaluate what
a \code{simList} would look like once it is converted to a repeatably digestible object.
}
\details{
See \code{\link[reproducible:.robustDigest]{reproducible::.robustDigest()}}.
This method strips out stuff from a \code{simList} class object that would make it otherwise not
reproducibly digestible between sessions, operating systems, or machines.
This will likely still not allow identical digest results across R versions.
}
\seealso{
\code{\link[reproducible:.robustDigest]{reproducible::.robustDigest()}}
}
\author{
Eliot McIntire
}
