% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-checking.R
\name{.runCodeChecks}
\alias{.runCodeChecks}
\title{Runs a series of code checks during \code{simInit()}}
\usage{
.runCodeChecks(sim, m, k, hadPrevMessage = FALSE)
}
\arguments{
\item{sim}{a \code{simList} object}

\item{m}{module name}

\item{hadPrevMessage}{}

\item{message}{rest of message}
}
\value{
NULL (invisibly); invoked for side effect of generating message with that starts with
\code{paste0(m, ":", message)}.
}
\description{
This uses custom tools and some optional tools in the \code{codetools} package to check for function
collisions with known, common function collisions (\code{raster::stack}, \code{raster::scale}).
All outputs will be sent as messages.
}
\keyword{internal}
