% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-checking.R
\name{.runCodeChecks}
\alias{.runCodeChecks}
\title{Runs a series of code checks during simInit}
\usage{
.runCodeChecks(sim, m, k, hadPrevMessage = FALSE)
}
\arguments{
\item{sim}{simList}

\item{m}{module name}

\item{hadPrevMessage}{}

\item{message}{rest of message}
}
\value{
A message with that starts with paste0(m, ":", message)
}
\description{
This uses codetools::codeCheck for function consistency, and
codetools::findGlobals to check for function collisions with known,
common function collisions (raster::stack, raster::scale)
}
\keyword{internal}
