% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-simInit.R
\name{.runModuleInputObjects}
\alias{.runModuleInputObjects}
\title{Run module's \code{.inputObjects}}
\usage{
.runModuleInputObjects(sim, m, objects, notOlderThan)
}
\description{
Run \code{.inputObjects()} from each module file from each module, one at a time,
and remove it from the \code{simList} so next module won't rerun it.
}
\keyword{internal}
