% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{saveSimList}
\alias{saveSimList}
\title{Save a whole \code{simList} object to disk}
\usage{
saveSimList(sim, filename, fileBackendToMem = TRUE,
  filebackedDir = NULL, envir, ...)
}
\arguments{
\item{sim}{Either a \code{simList} or a character string of the name
of a \code{simList} that can be found in \code{envir}. Using
a character string will assign that object name to the saved
\code{simList}, so when it is recovered it will be given that
name.}

\item{filename}{Character string with the path for saving \code{simList}}

\item{fileBackendToMem}{Logical. If there are file-backed \code{Raster}
objects, should they be converted to memory objects, or should
they be kept as file-backed rasters w.
or loaded into RAM and saved within the \code{.RData} file.
If \code{TRUE} (default), then the files will be copied to
\code{file.path(dirname(filename), "rasters")}.}

\item{filebackedDir}{Character string, indicating a path to save
the files to in file-backed objects (currently, only Raster* objects)}

\item{envir}{If \code{sim} is a character string, then this must be provided.
It is the environment where the object named \code{sim} can
be found.}

\item{...}{Passed to \code{save}, e.g., \code{compression}}
}
\value{
A saved \code{.RData} file in \code{filename} location.
}
\description{
Because of the environment slot, this is not quite as straightforward as
just saving the object. This also has option for file-backed Rasters.
}
\seealso{
zipSimList
}
