% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-simInit.R
\name{simInitAndSpades}
\alias{simInitAndSpades}
\title{Put simInit and spades together}
\usage{
simInitAndSpades(...)
}
\arguments{
\item{...}{Passed to simInit}
}
\description{
This may allow for more efficient Caching. This passes all
arguments to simInit, then the created \code{simList} is passed
to \code{spades}
}
