% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simList-accessors.R
\docType{methods}
\name{citation}
\alias{citation}
\alias{citation,.simList-method}
\alias{citation,character-method}
\alias{inputObjects}
\alias{inputObjects,.simList-method}
\alias{outputObjects}
\alias{outputObjects,.simList-method}
\alias{reqdPkgs}
\alias{reqdPkgs,.simList-method}
\alias{documentation}
\alias{documentation,.simList-method}
\title{Metadata accessors}
\usage{
citation(package, lib.loc = NULL, auto = NULL, module = character())

\S4method{citation}{.simList}(package, lib.loc = NULL, auto = NULL,
  module = character())

\S4method{citation}{character}(package, lib.loc = NULL, auto = NULL,
  module = character())

inputObjects(sim, module)

\S4method{inputObjects}{.simList}(sim, module)

outputObjects(sim, module)

\S4method{outputObjects}{.simList}(sim, module)

reqdPkgs(sim, module)

\S4method{reqdPkgs}{.simList}(sim, module)

documentation(sim, module)

\S4method{documentation}{.simList}(sim, module)
}
\arguments{
\item{package}{For compatibility with \code{\link[utils]{citation}}. This can be
a \code{simList} or a character string for a package name.}

\item{lib.loc}{a character vector with path names of \R libraries, or
    the directory containing the source for \code{package}, or
    \code{NULL}.  The default value of \code{NULL} corresponds to all
    libraries currently known.  If the default is used, the loaded
    packages are searched before the libraries.}

\item{auto}{a logical indicating whether the default citation
    auto-generated from the package \file{DESCRIPTION} metadata should
    be used or not, or \code{NULL} (default), indicating that a
    \file{CITATION} file is used if it exists, or an object of class
    \code{"\link{packageDescription}"} with package metadata (see
    below).}

\item{module}{Optional character string indicating which module params should come from.}

\item{sim}{A \code{simList} object from which to extract element(s) or
in which to replace element(s).}
}
\description{
These accessors extract the metadata for a module (if specified) or all modules
in a \code{simList} if not specified.
}
