% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-methods.R
\name{.getOption}
\alias{.getOption}
\title{A slightly modified version of getOption}
\usage{
.getOption(x)
}
\arguments{
\item{x}{a character string holding an option name.}
}
\description{
This can take x as a character string or as a function that returns a character string.
}
\keyword{internal}
