% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat.R
\name{SpaDEStestSetUpDirectories}
\alias{SpaDEStestSetUpDirectories}
\title{SpaDES test: Set up directories}
\usage{
SpaDEStestSetUpDirectories(
  testPaths = "testdata",
  copyModule = testthat::is_testing(),
  modules = NULL,
  teardownEnv = if (testthat::is_testing()) testthat::teardown_env(),
  tempDir = tempdir()
)
}
\arguments{
\item{testPaths}{file or directory paths within the \code{tests/testthat}
directory to add to the file list.
By default a test data directory is included with location \code{tests/testthat/testdata}.}

\item{copyModule}{logical. If TRUE, the RProject must be a module.
It will be copied to the temporary test directories before testing.}

\item{modules}{character. if \code{copyModule = TRUE}, copy these modules.
By default, only the RProject module is copied.
If other modules are on the list, they must also be located in the RProject parent directory.}

\item{teardownEnv}{environment. Optional. Environment to use for scoping.
The default for testing is the \code{testthat::teardown_env()}.}

\item{tempDir}{character. Optional. Path to location of temporary test directory.}
}
\value{
list of test paths
}
\description{
Set up a temporary directory structure. See \strong{Details}.
}
\details{
This function was designed to be called in a \code{tests/testthat/setup.R} file.

This function will create a temporary directory structure
that will be removed on test teardown.
By default, the RProject will be considered a module, and it is copied
to the "modules" sub-directory to be available to tests.
An R package "library" sub-directory will be initialized
and the initial library paths are restored on teardown.
}
