% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat.R
\name{SpaDEStestSetUpDirectories}
\alias{SpaDEStestSetUpDirectories}
\title{SpaDES test: Set up directories}
\usage{
SpaDEStestSetUpDirectories(
  testPaths = "testdata",
  modulePath = NULL,
  moduleRepos = NULL,
  require = NULL,
  copyModules = testthat::is_testing(),
  teardownEnv = if (testthat::is_testing()) testthat::teardown_env(),
  tempDir = tempdir(),
  ...
)
}
\arguments{
\item{testPaths}{file or directory paths within the \code{tests/testthat}
directory to add to the file list.
By default a test data directory is included with location \code{tests/testthat/testdata}.}

\item{modulePath}{character.
By default, it is assumed the R project is a single module to be tested.
Otherwise, provide a directory path (relative to the R project root)
that contains modules to be tested.
Ignored if \code{moduleRepos} is provided.}

\item{moduleRepos}{character. Github repository locations of modules to test.}

\item{require}{character. Additional R packages to require}

\item{copyModules}{logical.
If TRUE, local modules will be copied to the temporary test directories before testing.
Ignored if \code{moduleRepos} is provided.}

\item{teardownEnv}{environment. Optional. Environment to use for scoping.
The default for testing is the \code{testthat::teardown_env()}.}

\item{tempDir}{character. Optional. Path to location of temporary test directory.}

\item{...}{passed to \code{\link[SpaDES.project]{setupProject}}}
}
\value{
list of test paths
}
\description{
Set up a temporary directory structure. See \strong{Details}.
}
\details{
This function was designed to be called in a \code{tests/testthat/setup.R} file.

This function will create a temporary directory structure
that will be removed on test teardown.
\code{\link[SpaDES.project]{setupProject}} is called to initialize modules and R packages.
}
