% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat.R
\name{SpaDEStestSetUpDirectories}
\alias{SpaDEStestSetUpDirectories}
\title{SpaDES test: Set up directories}
\usage{
SpaDEStestSetUpDirectories(
  modulePath = NULL,
  modules = NULL,
  testPaths = "testdata",
  inputPath = getOption("spades.test.paths.inputs"),
  cachePath = getOption("spades.test.paths.cache"),
  packagePath = getOption("spades.test.paths.packages"),
  tempDir = tempdir(),
  teardownEnv = if (testthat::is_testing()) testthat::teardown_env()
)
}
\arguments{
\item{modulePath}{character.
By default, it is assumed that the working directory is a module directory.
Otherwise, provide a directory path (absolute or relative to the R project root)
that contains modules to be tested.
Set to NA to not copy any modules for testing.}

\item{modules}{character. Module(s) to copy for testing.
Defaults to the working directory module.
If \code{modulePath} is provided, all modules in this directory are included by default.}

\item{testPaths}{file or directory paths within the \code{tests/testthat}
directory to add to the file list.
By default a test data directory is included with location \code{tests/testthat/testdata}.}

\item{inputPath}{character. Optional alternative location of the input data directory.
Defaults to \code{getOption("spades.test.paths.inputs")}.
This allows users to speed up testing by allowing inputs to persist between test runs.}

\item{cachePath}{character. Optional alternative location of the cache directory.
Defaults to \code{getOption("spades.test.paths.cache")}.
This allows users to speed up testing by allowing cache to persist between test runs.}

\item{packagePath}{character. Optional alternative location of the R packages directory.
Defaults to \code{getOption("spades.test.paths.packages")}.
This allows users to speed up testing by allowing packages to persist between test runs.}

\item{tempDir}{character. Optional. Path to location of temporary test directory.}

\item{teardownEnv}{environment. Optional. Environment to use for scoping.
The default for testing is the \code{testthat::teardown_env()}.}
}
\value{
list of test paths
}
\description{
List test data directories and set up a temporary directory structure
that will be removed on test teardown. See \strong{Details}.
}
\details{
This function was designed to be called in a \code{tests/testthat/setup.R} file.

Module(s) will be copied to a temporary testing directory for testing.
}
