% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadInputs.R
\name{suppliedElsewhere}
\alias{suppliedElsewhere}
\title{Assess whether an object has or will be supplied from elsewhere}
\usage{
suppliedElsewhere(object, sim, where = c("sim", "user", "initEvent"))
}
\description{
When loading objects into a \code{simList}, especially during the
\code{simInit} call, and inside the \code{.inputObjects} functions of modules,
it is often useful to know if an object in question will or has been
by the user via the \code{inputs} or \code{objects} arguments, or by another
module's \code{.inputObjects} while preparing its expected inputs (via
\code{expectsInputs} in metadata), or if it will be supplied by another
module during its \code{"init"} event. In all these cases, it may not
be necessary for a given module to load any default value for its \code{expectsInputs}.
This function can be used as a check to determine whether the module needs
to proceed in getting and assigning its default value.
}
