% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.tagsByClass,simList-method}
\alias{.tagsByClass,simList-method}
\title{\code{.tagsByClass} for \code{simList} objects}
\usage{
\S4method{.tagsByClass}{simList}(object)
}
\arguments{
\item{object}{Any R object.}
}
\description{
See \code{\link[reproducible:tagsByClass]{.tagsByClass}}. Adds current \code{moduleName},
\code{eventType}, \code{eventTime}, and \verb{function:spades} as \code{userTags}.
}
\seealso{
\code{\link[reproducible:tagsByClass]{.tagsByClass}}
}
\author{
Eliot McIntire
}
