% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-methods.R
\name{updateList}
\alias{updateList}
\title{Update elements of a named list with elements of a second named list}
\usage{
updateList(x, y)
}
\arguments{
\item{x, y}{a named list}
}
\value{
A named list, with elements sorted by name.
The values of matching elements in list \code{y}
replace the values in list \code{x}.
}
\description{
Defunct. Use \code{\link[utils:modifyList]{utils::modifyList()}} (which can not handle NULL) or
\code{\link[Require:modifyList2]{Require::modifyList2()}} for case with >2 lists and can handle NULL lists.
}
\author{
Alex Chubaty
}
