% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-methods.R
\name{updateList}
\alias{updateList}
\title{Update elements of a named list with elements of a second named list}
\usage{
updateList(x, y)
}
\arguments{
\item{x}{a named list}

\item{y}{a named list}
}
\value{
A named list, with elements sorted by name.
         The values of matching elements in list \code{y}
         replace the values in list \code{x}.
}
\description{
Deprecated. Use \code{\link[utils]{modifyList}} (which can not handle NULL) or
\code{\link[Require]{modifyList2}} for case with >2 lists and can handle NULL lists.
}
\examples{
L1 <- list(a = "hst", b = NA_character_, c = 43)
L2 <- list(a = "gst", c = 42, d = list(letters))
updateList(L1, L2)

updateList(L1, NULL)
updateList(L1)
updateList(y = L2)
updateList(NULL, L2)
updateList(NULL, NULL) # should return empty list

}
\author{
Alex Chubaty
}
