% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-template.R
\name{use_gha}
\alias{use_gha}
\title{Use GitHub actions for automated module checking}
\usage{
use_gha(name, path)
}
\arguments{
\item{name}{module name}

\item{path}{module path}
}
\value{
Invoked for its side effect of creating new GitHub Actions workflow files.
}
\description{
See corresponding vignette for more information.
}
