% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadInputs.R
\name{writeInputsOnDisk}
\alias{writeInputsOnDisk}
\title{Write module inputs on disk}
\usage{
writeInputsOnDisk(x, filename, ...)
}
\arguments{
\item{x}{An object of postProcessing. See individual methods.}

\item{filename}{The filename to save the output object to disk (a \code{Raster*} or
\code{Spatial*} object)}

\item{...}{Passed to \code{\link[raster]{writeRaster}}, such as \code{datatype},
and \code{\link[raster]{shapefile}}}
}
\description{
Can be used to write prepared inputs on disk.
}
\author{
Eliot McIntire

Jean Marchal
}
